Introduce implementation information to the initialize phase (#192)
* Introduce implementation information to the initialize phase

This introduces a way for Clients and Agents to provide information about themselves to either party.

This can be used to show users which implementation and it's version to users, or be used for logging/metrics purposes to figure out the source of usage.

This is being introduced as an optional field for now, so it is backwards compatible, but it seems like a likely candidate to be made into a required field for v2 so both sides can count on this information being available.

Copying roughly the exact same pattern as MCP here: https://modelcontextprotocol.io/specification/2025-06-18/basic/lifecycle#initialization

* Start a v2 document

* Fix clippy