Clean up JSON Schema (#252)
* Clean up JSON Schema

We made some changes in the past to inline a bunch of $refs to make
things easier for sdk generation. However, this came with downsides of
less distinct types.

This should have a more compatible version, that still preserves the
non-inlined types.

At least on the TypeScript side this has vastly improved the generated
types.

* Inline json rpc stuff again

* Clean up defaults

* Remove unneded serde renames

* Replace bool schemas

Helps with compatibility of more JSON Schema variants

* Bring back default change as it has weird effects with anyof

* Fix references in generated docs with the allOf change

* Better merge properties in the docs for types that extend refs