feat!: Provide clearer schema for _meta properties (#290)
This defines the `_meta` properties in the schema as intended and used, which is always an object of key/value pairs, with string keys and arbitrary values.

While this is how everyone is using them, it became clear in code generation that the types weren't quite matching up to the expected usage. This should alleviate some extra checks on the implementor side.