feat(rust-only): More convenient builder method params (#313)
This is a Rust-only feature, but for optinal fields, the caller can now pass in either the value itself, or an option of that value, and we'll take care of the rest. It provides a lot of flexibility when constructing these values, as it turned out our callers often had options, and the `(mut self, T) -> Self` convention made this a real pain if you had to optionally call the builder method. Now you can call it either way and it reads really nice!